+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # ChangeLog #
 
+## 0.2.0 - Feb 23, 2014
+* Fixed curry expression args not always being an array.
+* Better error for curry expression.
+
 ## 0.1.0 - Feb 19, 2014
 * Allow self unpack to be any of: array pattern, as pattern, object pattern,
   identifier pattern.
@@ -7,7 +11,7 @@
 ## 0.0.1 - Feb 3, 2014
 * Removed use of custom ParserState for parsing
 * Removed support for unicode `\\` escaped sequences in identifiers which was
-  always lexed to the `\\` puctuator anyways.
+  always lexed to the `\\` punctuator anyways.
 
 ## 0.0.0 - Feb 3, 2014
 * Fork from main Khepri codebase.